requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.2.3
django-admin-interface==0.16.3
django-ckeditor==6.0.0
django-colorfield==0.4.1
django-crispy-forms==1.11.2
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-heroku==0.3.1
django-js-asset==1.2.2
gunicorn==20.1.0
idna==2.10
numpy==1.20.2
pandas==1.2.4
Pillow==8.2.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
sentry-sdk==1.0.0
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.2.0
